1，履带电机pid算法优化，主要是应对履带电机打滑的问题（王仁飞） 2，风机pid算法优化，主要是风机启动慢，导致pid疯狂的增加或减少占空比，导致风机忽大忽小（王仁飞） 3，优化陀螺仪驱动，补偿零漂的Gy计算Pitch改变（王仁飞） 4，修复GPRS通信时的环形buff头尾指针更新有误（李坤鹏） 5，新增风机完全停止计时器，等风机完全停住时才记录空闲时气压（李坤鹏）

+++ YKC/BoardSupport.h
@@ -210,7 +210,7 @@ typedef struct {
 
 #ifdef _LIB_MATH_C_
 const I2C_BSP cnst_I2C_BSP[I2C_COMM_NUM] = {
-		{GPIOB, GPIO_PIN_8, GPIOB, GPIO_PIN_9}
+		{(uint32)GPIOB, GPIO_PIN_8, (uint32)GPIOB, GPIO_PIN_9}
 };
 #else
 extern const I2C_BSP cnst_I2C_BSP[I2C_COMM_NUM];
